feat: 能力値Float化・決定論的成長システム実装
- Stats: 全フィールドをInt→Floatに変更、effectiveXxx(Int切捨)プロパティ追加
- GrowthRate: Int%(確率成長)→Float固定値加算(決定論的成長)に変更
- StatGrowth: 新規データクラス（レベルアップUI用Int変化量）
- GameUnit: applyLevelUp()を決定論的成長に変更、rollGrowth()削除
- SPD成長: 全ユニット共通0.20f/Lv（FFT方式、5Lvで+1）
- DamageCalc/Screen: stats.xxx→stats.effectiveXxx
- SaveManager: SAVE_VERSION=2、v1→v2マイグレーション対応
- MapLoader: Float対応、敵レベル生成の決定論化
- units.json: GrowthRate値をFloat化
- 全14テストファイル更新、165テスト全件通過
- 仕様書: 10-growth-system.md新規作成、09/README更新

+++ core/src/main/kotlin/com/tacticsflame/model/unit/GrowthRate.kt
@@ -1,25 +1,29 @@
 package com.tacticsflame.model.unit
 
 /**
- * レベルアップ時の成長率を表すデータクラス
- * 各値は0〜100のパーセンテージ
+ * レベルアップ時の確定成長値を表すデータクラス
  *
- * @property hp HP成長率（%）
- * @property str 力成長率（%）
- * @property mag 魔力成長率（%）
- * @property skl 技成長率（%）
- * @property spd 速さ成長率（%）
- * @property lck 幸運成長率（%）
- * @property def 守備成長率（%）
- * @property res 魔防成長率（%）
+ * 各値はレベルアップごとに加算される固定の Float 値。
+ * 例: hp = 0.70f → 毎レベルアップで HP に +0.70 加算。
+ *
+ * SPD は原則として全ユニット共通の値（0.20）を設定する（FFT方式）。
+ *
+ * @property hp HP成長値
+ * @property str 力成長値
+ * @property mag 魔力成長値
+ * @property skl 技成長値
+ * @property spd 速さ成長値（全ユニット共通 0.20 推奨）
+ * @property lck 幸運成長値
+ * @property def 守備成長値
+ * @property res 魔防成長値
  */
 data class GrowthRate(
-    val hp: Int = 0,
-    val str: Int = 0,
-    val mag: Int = 0,
-    val skl: Int = 0,
-    val spd: Int = 0,
-    val lck: Int = 0,
-    val def: Int = 0,
-    val res: Int = 0
+    val hp: Float = 0f,
+    val str: Float = 0f,
+    val mag: Float = 0f,
+    val skl: Float = 0f,
+    val spd: Float = 0f,
+    val lck: Float = 0f,
+    val def: Float = 0f,
+    val res: Float = 0f
 )
